+++ packages/router-control/src/interface.ts
@@ -24,6 +24,7 @@ export interface RoutesBaseProps extends Omit<RouteObject, 'children'> {
 
 // json文件格式
 export interface RoutersJSON extends Omit<RoutesBaseProps, "routes"> {
+  /** 组件地址 如果是 403/404/500 的页面直接写 403/404/500 就可以了，内部直接做转化*/
   component?: string
   /** 子集 路由 */
   routes?: RoutersJSON[];