+++ examples/base/.kktrc.ts
@@ -1,18 +1,18 @@
-import path from 'path';
-import { MockerAPIOptions } from 'kkt';
-import lessModules from '@kkt/less-modules';
-import rawModules from '@kkt/raw-modules';
-import scopePluginOptions from '@kkt/scope-plugin-options';
-import pkg from './package.json';
-import defaultConfig from '@uiw-admin/config';
+import path from 'path'
+import { MockerAPIOptions } from 'kkt'
+import lessModules from '@kkt/less-modules'
+import rawModules from '@kkt/raw-modules'
+import scopePluginOptions from '@kkt/scope-plugin-options'
+import pkg from './package.json'
+import defaultConfig from '@uiw-admin/config'
 import {
   RematchWebpackPlugin, WidgetsWebpackPlugin,
   RoutesWebpackPlugin
-} from '@uiw-admin/plugins';
+} from '@uiw-admin/plugins'
 
 export default defaultConfig({
   define: {
-    VERSION: JSON.stringify(pkg.version),
+    VERSION: JSON.stringify(pkg.version)
     // BASE_NAME: "/uiw"
   },
   plugins: [
@@ -25,11 +25,11 @@ export default defaultConfig({
     rawModules,
     {
       loader: scopePluginOptions,
-      options: { allowedFiles: [path.resolve(process.cwd(), 'README.md')] },
+      options: { allowedFiles: [path.resolve(process.cwd(), 'README.md')] }
     },
-    lessModules,
-  ],
-});
+    lessModules
+  ]
+})
 export const proxySetup = (): MockerAPIOptions => {
   /**
    * mocker-api that creates mocks for REST APIs.
@@ -41,6 +41,6 @@ export const proxySetup = (): MockerAPIOptions => {
     /**
      * https://github.com/jaywcjlove/mocker-api/tree/96c2eb94694571e0e3003e6ad9ce1c809499f577#options
      */
-    option: {},
-  };
-};
+    option: {}
+  }
+}