+++ examples/base/config/routes.js
@@ -88,6 +88,6 @@ const routes = [
       },
     ],
   },
-];
+]
 
-export default routes;
+export default routes