+++ examples/base/src/index.tsx
@@ -1,24 +1,24 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import Control from '@uiw-admin/router-control';
-import { SWRConfig } from 'swr';
-import { request } from '@uiw-admin/utils';
-import '@uiw/reset.css';
-import './index.css';
+import React from 'react'
+import ReactDOM from 'react-dom'
+import Control from '@uiw-admin/router-control'
+import { SWRConfig } from 'swr'
+import { request } from '@uiw-admin/utils'
+import '@uiw/reset.css'
+import './index.css'
 
 ReactDOM.render(
   <SWRConfig
     value={{
       // revalidateOnFocus: false,
       fetcher: (resource, init) => {
-        return request(resource, init);
+        return request(resource, init)
       },
-      provider: () => new Map(),
+      provider: () => new Map()
     }}
   >
     <Control
       routeType="hash"
     />
   </SWRConfig>,
-  document.getElementById('root'),
-);
+  document.getElementById('root')
+)