+++ examples/base/src/layouts/BasicLayout.tsx
@@ -1,10 +1,10 @@
-import React from 'react';
-import BasicLayout from '@uiw-admin/basic-layouts';
-import { Outlet } from 'react-router-dom';
-import { RoutersProps } from '@uiw-admin/router-control';
-import { Badge, Icon } from 'uiw';
+import React from 'react'
+import BasicLayout from '@uiw-admin/basic-layouts'
+import { Outlet } from 'react-router-dom'
+import { RoutersProps } from '@uiw-admin/router-control'
+import { Badge, Icon } from 'uiw'
 
-import useSWR from 'swr';
+import useSWR from 'swr'
 
 // import LayoutTabs from "@uiw-admin/layout-tabs"
 // import Auth from "@uiw-admin/authorized"
@@ -13,19 +13,19 @@ interface BasicLayoutProps {
   routes: RoutersProps[];
 }
 
-function BasicLayoutScreen(props: BasicLayoutProps = { routes: [] }) {
-  const { routes } = props;
+function BasicLayoutScreen (props: BasicLayoutProps = { routes: [] }) {
+  const { routes } = props
   const { mutate } = useSWR(['/api/reloadAuth', { method: 'POST' }], {
     revalidateOnMount: false,
     revalidateOnFocus: false,
     onSuccess: (data) => {
       if (data && data.code === 200) {
-        sessionStorage.setItem('token', data.token);
-        sessionStorage.setItem('auth', JSON.stringify(data.authList || []));
-        window.location.reload();
+        sessionStorage.setItem('token', data.token)
+        sessionStorage.setItem('auth', JSON.stringify(data.authList || []))
+        window.location.reload()
       }
-    },
-  });
+    }
+  })
 
   const basicLayoutProps = {
     onReloadAuth: async () => mutate(),
@@ -35,13 +35,13 @@ function BasicLayoutScreen(props: BasicLayoutProps = { routes: [] }) {
       {
         title: '欢迎来到uiw',
         icon: 'smile',
-        onClick: () => {},
+        onClick: () => null
       },
       {
         title: '修改密码',
         icon: 'setting',
-        onClick: () => {},
-      },
+        onClick: () => null
+      }
     ],
     profile: {
       avatar: require('../assets/head.png'),
@@ -51,9 +51,9 @@ function BasicLayoutScreen(props: BasicLayoutProps = { routes: [] }) {
             <Icon type="bell" color="#343a40" style={{ fontSize: 20 }} />
           </Badge>
         </div>
-      ),
-    },
-  };
+      )
+    }
+  }
 
   // 验证是否登录的方式
   // 1. 使用 Auth 组件
@@ -71,6 +71,6 @@ function BasicLayoutScreen(props: BasicLayoutProps = { routes: [] }) {
       <Outlet />
       {/* <LayoutTabs routes={routes || []} /> */}
     </BasicLayout>
-  );
+  )
 }
-export default BasicLayoutScreen;
+export default BasicLayoutScreen