+++ examples/base/src/models/login.ts
@@ -1,5 +1,5 @@
-import { ModelDefault } from '@uiw-admin/models';
-import { history } from '@uiw-admin/router-control';
+import { ModelDefault } from '@uiw-admin/models'
+import { history } from '@uiw-admin/router-control'
 
 export interface LoginState {
   token?: string;
@@ -12,25 +12,25 @@ const login: ModelDefault = {
   name: 'login',
   state: {
     userData: null,
-    token: null,
+    token: null
   },
   reducers: {
     updateState: (state: any, payload: LoginState) => ({
       ...state,
-      ...payload,
-    }),
+      ...payload
+    })
   },
   effects: (dispatch: any) => ({
-    async submit(payload) {
-      dispatch.login.updateState({ token: '测试2' });
-      sessionStorage.setItem('auth', JSON.stringify(['/home', '/dac']));
-      history.push('/home');
+    async submit () {
+      dispatch.login.updateState({ token: '测试2' })
+      sessionStorage.setItem('auth', JSON.stringify(['/home', '/dac']))
+      history.push('/home')
       // this.updateState()
       // await login({ username: 'test', password: 'www' });
       // dispatch.sharks.increment(payload)
       // `dispatch.s` will suggest `sharks`
-    },
-  }),
-};
+    }
+  })
+}
 
-export default login;
+export default login